requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
python-dotenv
google-generativeai
plotly